+++ packages/adp-tooling/src/preview/adp-preview.ts
@@ -22,6 +22,8 @@ import {
     addXmlFragment,
     isAddAnnotationChange,
     isAddXMLChange,
+    isCodeExtChange,
+    addControllerExtension,
     moduleNameContentMap,
     tryFixChange
 } from './change-handler';
@@ -259,6 +261,15 @@ export class AdpPreview {
             case 'write':
                 if (isAddXMLChange(change)) {
                     addXmlFragment(this.util.getProject().getSourcePath(), change, fs, logger, additionalChangeInfo);
+                }
+                if (isCodeExtChange(change)) {
+                    await addControllerExtension(
+                        this.util.getProject().getRootPath(),
+                        this.util.getProject().getSourcePath(),
+                        change,
+                        fs,
+                        logger
+                    );
                 }
                 if (isAddAnnotationChange(change)) {
                     await addAnnotationFile(