+++ packages/preview-middleware-client/test/unit/adp/controllers/AddTableColumnFragments.controller.test.ts
@@ -16,6 +16,7 @@ import type ManagedObject from 'sap/ui/base/ManagedObject';
 import AddTableColumnFragments from 'open/ux/preview/client/adp/controllers/AddTableColumnFragments.controller';
 import SimpleForm from 'sap/ui/layout/form';
 import Control from 'sap/ui/core/Control';
+import * as adpUtils from 'open/ux/preview/client/adp/utils';
 import { getAdditionalChangeInfo } from '../../../../src/utils/additional-change-info';
 import { FlexChange } from 'open/ux/preview/client/flp/common';
 
@@ -220,8 +221,6 @@ describe('AddTableColumnsFragments controller', () => {
                 mockFormInput(true, '', ValueState.Success)
             ] as unknown as Control[]);
 
-            addFragment.checkForExistingChange = jest.fn().mockReturnValue(false);
-
             addFragment.onColumnFragmentNameInputChange(event as unknown as Event);
             addFragment.onCellFragmentNameInputChange(event as unknown as Event);
             expect(mocks.setValueStateMock).toHaveBeenCalledTimes(2);
@@ -233,14 +232,13 @@ describe('AddTableColumnsFragments controller', () => {
 
         describe('duplicate fragment names', () => {
             test('sets error when column and cell fragments have the same name', () => {
+                jest.spyOn(adpUtils, 'checkForExistingChange').mockReturnValue(false);
                 const event = mockInputEvent('Name1');
 
                 createDialog([
                     mockFormInput(true, 'Name1', [ValueState.Success, ValueState.Error]),
                     mockFormInput(true, 'Name1', [ValueState.Success, ValueState.Error])
                 ] as unknown as Control[]);
-
-                addFragment.checkForExistingChange = jest.fn().mockReturnValue(false);
 
                 addFragment.onColumnFragmentNameInputChange(event as unknown as Event);
 
@@ -265,8 +263,6 @@ describe('AddTableColumnsFragments controller', () => {
                     mockFormInput(true, 'Delete', ValueState.Error, errorMsg)
                 ] as unknown as Control[]);
 
-                addFragment.checkForExistingChange = jest.fn().mockReturnValue(false);
-
                 addFragment.onColumnFragmentNameInputChange(event as unknown as Event);
 
                 expect(mocks.setValueStateMock).toHaveBeenCalledTimes(1);
@@ -275,14 +271,13 @@ describe('AddTableColumnsFragments controller', () => {
             });
 
             test('clears errors on value change', () => {
+                jest.spyOn(adpUtils, 'checkForExistingChange').mockReturnValue(false);
                 const event = mockInputEvent('Name2');
 
                 createDialog([
                     mockFormInput(true, 'Name2', [ValueState.Error, ValueState.Success], ['Duplicate name', '']),
                     mockFormInput(true, 'Delete', [ValueState.Error, ValueState.Success], ['Duplicate name', ''])
                 ] as unknown as Control[]);
-
-                addFragment.checkForExistingChange = jest.fn().mockReturnValue(false);
 
                 addFragment.onColumnFragmentNameInputChange(event as unknown as Event);
 
@@ -301,8 +296,6 @@ describe('AddTableColumnsFragments controller', () => {
                 mockFormInput(true, 'Name2', ValueState.Success)
             ] as unknown as Control[]);
 
-            addFragment.checkForExistingChange = jest.fn().mockReturnValue(false);
-
             addFragment.onColumnFragmentNameInputChange(event as unknown as Event);
 
             expect(mocks.setValueStateMock).toHaveBeenCalledWith(ValueState.None);
@@ -316,8 +309,6 @@ describe('AddTableColumnsFragments controller', () => {
                 mockFormInput(true, ['Name1', 'Share2$5!'], [ValueState.Success, ValueState.Error]),
                 mockFormInput(true, 'Name2', ValueState.Success)
             ] as unknown as Control[]);
-
-            addFragment.checkForExistingChange = jest.fn().mockReturnValue(false);
 
             addFragment.onColumnFragmentNameInputChange(event as unknown as Event);
 
@@ -335,8 +326,6 @@ describe('AddTableColumnsFragments controller', () => {
                 mockFormInput(true, ['Name1', 'Name '], [ValueState.Success, ValueState.Error]),
                 mockFormInput(true, 'Name2', ValueState.Success)
             ] as unknown as Control[]);
-
-            addFragment.checkForExistingChange = jest.fn().mockReturnValue(false);
 
             addFragment.onColumnFragmentNameInputChange(event as unknown as Event);
 
@@ -357,8 +346,6 @@ describe('AddTableColumnsFragments controller', () => {
                 mockFormInput(true, 'Name2', ValueState.Success)
             ] as unknown as Control[]);
 
-            addFragment.checkForExistingChange = jest.fn().mockReturnValue(false);
-
             addFragment.onColumnFragmentNameInputChange(event as unknown as Event);
 
             expect(mocks.setValueStateMock).toHaveBeenCalledWith(ValueState.Error);
@@ -369,6 +356,7 @@ describe('AddTableColumnsFragments controller', () => {
         });
 
         test('does not crash if composite command exists in command stack', () => {
+            jest.spyOn(adpUtils, 'checkForExistingChange').mockReturnValue(false);
             const rtaMock = new RuntimeAuthoringMock({} as RTAOptions);
 
             const command = {
@@ -392,6 +380,7 @@ describe('AddTableColumnsFragments controller', () => {
         });
 
         test('sets error when the fragment name already exists in command stack', () => {
+            jest.spyOn(adpUtils, 'checkForExistingChange').mockReturnValue(true);
             const rtaMock = new RuntimeAuthoringMock({} as RTAOptions);
             const change = {
                 content: {
@@ -421,6 +410,7 @@ describe('AddTableColumnsFragments controller', () => {
         });
 
         test('sets error when the fragment name already exists in command stack (command is "composite")', () => {
+            jest.spyOn(adpUtils, 'checkForExistingChange').mockReturnValue(true);
             const rtaMock = new RuntimeAuthoringMock({} as RTAOptions);
             const change = {
                 content: {
@@ -454,6 +444,7 @@ describe('AddTableColumnsFragments controller', () => {
         });
 
         test('sets create button to true when the fragment name is valid', () => {
+            jest.spyOn(adpUtils, 'checkForExistingChange').mockReturnValue(false);
             const rtaMock = new RuntimeAuthoringMock({} as RTAOptions);
             rtaMock.getCommandStack.mockReturnValue({
                 getCommands: jest.fn().mockReturnValue([])