+++ packages/ui-service-inquirer/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sap-ux/ui-service-inquirer
 
+## 0.1.1
+
+### Patch Changes
+
+-   Updated dependencies [66b88e1]
+    -   @sap-ux/inquirer-common@0.7.1
+    -   @sap-ux/odata-service-inquirer@2.4.1
+    -   @sap-ux/abap-deploy-config-inquirer@1.4.1
+    -   @sap-ux/telemetry@0.6.1
+    -   @sap-ux/axios-extension@1.21.0
+
 ## 0.1.0
 
 ### Minor Changes